change progress data type

+++ app/src/main/java/com/belajarkuy/app/data/model/CompetencyResponse.kt
@@ -27,5 +27,5 @@ data class CompetencyItem(
 	val subject: String,
 
 	@field:SerializedName("progress")
-	val progress: Int
+	val progress: Float
 ) : Parcelable
